Add files via upload
For use on a gaming PC that doubles as miner. Schedule this task to run in the background or once a minute and and it will check if GPU usage is 0 and, if it is, fire up your miner. If you then start your game, the miner will stop running.